api/tests: move rule and rule-section tests to own file
